# Console logging optimizations