… src/WebJobs.Script.WebHost/Diagnostics/LinuxContainerEventGenerator.cs
… src/WebJobs.Script.WebHost/WebHostServiceCollectionExtensions.cs
… src/WebJobs.Script/Config/ConsoleLoggingOptionsSetup.cs
… src/WebJobs.Script/Environment/EnvironmentSettingNames.cs
… test/WebJobs.Script.Tests/Configuration/ConsoleLoggingOptionsSetupTests.cs
… test/WebJobs.Script.Tests/Diagnostics/LinuxContainerEventGeneratorWithConsoleOutputTests.cs