Resolves #9228

For Linux Consumption, system logs go through standard out. Analysis shows that there is a significant cost associated with this code path, and the cost can be reduced by:
1. Moving the Console.WriteLine call to a background thread
2. Reducing the total number of Console.WriteLine calls by appending log lines together first.

This PR implements both of these improvements, but they can be controlled with the following environment variables:
- CONSOLE_LOGGING_BUFFER_SIZE - this controls the size of the log buffer (how many log lines), defaults to 1000, set to 0 to disable, set to -1 for unbounded buffer size.
- CONSOLE_LOGGING_BUFFER_BATCHED - when enabled, uses a string builder to write many log lines to stdout with a single Console.WriteLine call, defaults to enabled, set to zero to disable the batching.

Naming for these variables was based on the existing CONSOLE_LOGGING_DISABLED setting for consistency.

## Performance Test Results

All tests run with Python 3.9 function app running hello world,  median result from 3 runs.

### ACI (single instance, private network)

Baseline:
```
Bombarding http://py39-aci-noai-consolebuffer-baseline.app:80/api/helloworld for 1m0s using 30 connection(s)
Statistics        Avg      Stdev        Max
  Reqs/sec        94.88      63.33     949.95
  Latency      315.32ms    46.99ms   611.99ms
  Latency Distribution
     50%   315.00ms
     75%   343.34ms
     90%   371.79ms
     95%   389.36ms
     99%   450.89ms
  HTTP codes:
    1xx - 0, 2xx - 5715, 3xx - 0, 4xx - 0, 5xx - 0
    others - 0
  Throughput:    27.19KB/s
```

Buffer 1000, batched:
```
Bombarding http://py39-aci-noai-consolebuffer-buffer1000-batched.app:80/api/helloworld for 1m0s using 30 connection(s)
Statistics        Avg      Stdev        Max
  Reqs/sec       233.86     180.42    1526.60
  Latency      128.20ms    19.10ms   269.31ms
  Latency Distribution
     50%   129.00ms
     75%   143.78ms
     90%   157.98ms
     95%   167.23ms
     99%   188.00ms
  HTTP codes:
    1xx - 0, 2xx - 14050, 3xx - 0, 4xx - 0, 5xx - 0
    others - 0
  Throughput:    69.20KB/s
```

### Legion (single instance, private network)

Baseline:
```
Bombarding http://py39-noai-mesh-perfcollect-consolebuffer-baseline.app:80/api/helloworld for 1m0s using 30 connection(s)
Statistics        Avg      Stdev        Max
  Reqs/sec       342.15     233.37    1578.96
  Latency       88.11ms    64.55ms      0.98s
  Latency Distribution
     50%    73.80ms
     75%    83.00ms
     90%    93.00ms
     95%   102.52ms
     99%   693.74ms
  HTTP codes:
    1xx - 0, 2xx - 20538, 3xx - 0, 4xx - 0, 5xx - 0
    others - 0
  Throughput:   101.72KB/s
```
Buffer 1000, batched:
```
Bombarding http://py39-noai-mesh-perfcollect-consolebuffer-buffer1000-batched.app:80/api/helloworld for 1m0s using 30 connection(s)
Statistics        Avg      Stdev        Max
  Reqs/sec       393.31     248.12    1650.02
  Latency       76.23ms    33.30ms      0.86s
  Latency Distribution
     50%    71.22ms
     75%    81.00ms
     90%    90.00ms
     95%    97.29ms
     99%   233.16ms
  HTTP codes:
    1xx - 0, 2xx - 23628, 3xx - 0, 4xx - 0, 5xx - 0
    others - 0
  Throughput:   121.39KB/s
```

### Linux Consumption, EUAP (burst scale to 30 instances, tested over public internet)

1000 concurrent clients, burst scale test
```
Baseline: 1,183 RPS (average over 5 minutes)
Buffer 1000, batched: 1,603 RPS (average over 5 minutes)
```

### Test results summary
All runs with this PR had better results, though there is significant variability. I am not 100% sure about why my EUAP tests did not see the same gains as the single instance tests I ran on ACI, but I suspect that network travel time is a factor here, because the EUAP tests were conducted over the public internet. We'll wait to see whether we get similar numbers as part of the official release. Console.WriteLine performance on Legion is not a major issue, so the gains are smaller there, but still worthwhile.

## Functional testing checklist
* [x] Test in Linux Consumption EUAP and confirm that system logs appear in kusto as expected, with expected volume (i.e. requests completed during the test matches number of function completion events
* [x] Test in Legion and confirm system logs appear in kusto as expected

## Pull request checklist

* [x] My changes **do not** require documentation changes
    * [ ] Otherwise: Documentation issue linked to PR
* [ ] My changes **should not** be added to the release notes for the next release
    * [ ] Otherwise: I've added my notes to `release_notes.md`
* [x] My changes **do not** need to be backported to a previous version
    * [ ] Otherwise: Backport tracked by issue/PR #issue_or_pr
* [x] My changes **do not** require diagnostic events changes
    * Otherwise: I have added/updated all related diagnostic events and their documentation (Documentation issue linked to PR)
* [x] I have added all required tests (Unit tests, E2E tests)