@paulbatum 2023-05-09
I did some more testing based on the conversation about the fact that we occasionally have to block on a Task.Result due to the choice of using channels. I coded up some alternative versions:
1. BlockingCollection - similar what we do for linux dedicated, but with a bounded size.
2. Lock + Queue + Monitor for synchronization. This is based on the .NET console logging implementation: https://github.com/dotnet/runtime/blob/main/src/libraries/Microsoft.Extensions.Logging.Console/src/ConsoleLoggerProcessor.cs

Neither of these compared favorably to the current implementation in this PR - all results below are the median of 3 runs. These tests were performed in ACI this time around, so the results are different to the previous round that I shared.

Baseline:
```
Bombarding http://py39-aci-noai-consolebuffer-baseline.app:80/api/helloworld for 1m0s using 30 connection(s)
Statistics        Avg      Stdev        Max
  Reqs/sec       103.53      74.82     847.87
  Latency      289.34ms    41.55ms   490.04ms
  Latency Distribution
     50%   289.47ms
     75%   316.62ms
     90%   341.94ms
     95%   359.11ms
     99%   397.00ms
  HTTP codes:
    1xx - 0, 2xx - 6236, 3xx - 0, 4xx - 0, 5xx - 0
    others - 0
  Throughput:    29.63KB/s
```

Blocking Collection:
```
Bombarding http://py39-noai-mesh-perfcollect-consolebuffer-blocking8000.app:80/api/helloworld for 1m0s using 30 connection(s)
Statistics        Avg      Stdev        Max
  Reqs/sec        18.06     131.03    1499.98
  Latency         1.67s      2.36s     10.03s
  Latency Distribution
     50%    31.00ms
     75%    51.00ms
     90%     10.02s
     95%     10.02s
     99%     10.03s
  HTTP codes:
    1xx - 0, 2xx - 921, 3xx - 0, 4xx - 0, 5xx - 0
    others - 180
  Errors:
    the server closed connection before returning the first response byte. Make sure the server returns 'Connection: close' response header before closing the connection - 180
  Throughput:     6.31KB/s
```

Lock + Queue + Monitor:
```
Bombarding http://py39-aci-noai-consolebuffer-monitor8000.app:80/api/helloworld for 1m0s using 30 connection(s)
Statistics        Avg      Stdev        Max
  Reqs/sec       118.50     106.08    1525.66
  Latency      252.98ms    44.96ms   478.40ms
  Latency Distribution
     50%   253.10ms
     75%   279.24ms
     90%   308.81ms
     95%   328.65ms
     99%   377.51ms
  HTTP codes:
    1xx - 0, 2xx - 7125, 3xx - 0, 4xx - 0, 5xx - 0
    others - 0
  Throughput:    34.25KB/s
```

Channel with buffer size of 8000, batched (current PR):
```
Bombarding http://py39-aci-noai-consolebuffer-channel8000-batched.app:80/api/helloworld for 1m0s using 30 connection(s)
Statistics        Avg      Stdev        Max
  Reqs/sec       146.76     120.72    1299.97
  Latency      204.24ms    34.40ms   365.00ms
  Latency Distribution
     50%   205.38ms
     75%   230.00ms
     90%   252.93ms
     95%   268.00ms
     99%   298.22ms
  HTTP codes:
    1xx - 0, 2xx - 8821, 3xx - 0, 4xx - 0, 5xx - 0
    others - 0
  Throughput:    43.56KB/s
```

The BlockingCollection results are particularly awful, and this matches up with some very inconsistent results I got when testing BlockingCollection outside of the context of the functions code base. Specifically the problems arise when it has a fixed bound that is being reached. Our current Linux Dedicated implementation does not set a bound on the collection, so it doesn't run into this problem.

I think the Channel based implementation is still the best option.