@paulbatum 2023-05-15
I've ran some tests with large load and very small buffer sizes to determine how the buffer-full codepath behaves, and its not good - while there is no deadlock condition, the app shows signs of threadpool starvation, probably due to doing Task.Result on a sync codepath. 

```
Bombarding http://py39-perfcollect-channel100-batched.app:80/api/helloworld for 1m0s using 100 connection(s)
Statistics        Avg      Stdev        Max
  Reqs/sec        26.29     158.67    2818.03
  Latency         3.84s      3.02s     10.04s
  Latency Distribution
     50%   267.03ms
     75%     10.01s
     90%     10.02s
     95%     10.02s
     99%     10.03s
  HTTP codes:
    1xx - 0, 2xx - 1066, 3xx - 0, 4xx - 0, 5xx - 0
    others - 600
  Errors:
    the server closed connection before returning the first response byte. Make sure the server returns 'Connection: close' response header before closing the connection - 600
  Throughput:     9.48KB/s
```

Now with the planned buffer size of 8000 I don't think this condition is likely to be hit, but I consider this a deal breaker for the current implementation. I will do more testing and find an alternative, and I'll re-request reviews once I've pushed an update.