@jviau 2023-05-15
> > Do you have plans to flush out the buffer on process shutdown? If not, we will potentially lose console writes.
> 
> At first I was skeptical about how necessary this is, given that the linux dedicated implemention already buffers and doesn't do a flush, and we don't have any reports of lost logs. But on second thought, its reasonably cheap to implement by relying on the DI behavior of calling dispose on all registered services that implement IDisposable, so I'll push an update that does this.

We don't have any reports of lost logs? In my (limited) experience of diagnosing some function app issues, I have had several instances of what I chalked up to lost logs (but could not prove). These were all around a function app starting up, then being stopped